All required commands exist.
By required that means setting up the simulation, taking steps, adding and removing pivots and masses.  Listing out the masses and pivots.  Setting the height, length and depth of the plane.  Clearing the area.

Currently, does not support a continuous server side simulation auto step each tick, that has to be done on the server with commands.